refactor(emitter): split oversized files to satisfy §19 2000-line limit
Split `bindings_patterns.rs` (2764 lines) into `bindings_patterns.rs` +
`bindings_param_patterns.rs`, and split `core/tests.rs` (3041 lines) into
`core/tests/{mod,cjs_module_exports,esm_class_namespace,cjs_namespace_alias}.rs`.
Also makes cross-file private helpers `pub(super)` to fix visibility errors.
