+++ crates/tsz-emitter/src/emitter/es5/mod.rs
@@ -7,6 +7,7 @@ mod bindings;
 mod bindings_assignment;
 mod bindings_disposable_names;
 mod bindings_for_of;
+mod bindings_param_patterns;
 mod bindings_patterns;
 mod bindings_read;
 mod helpers;
